+++ frontend/app/src/main/java/com/markoala/tomoandroid/ui/main/friends/AddFriendsScreen.kt
@@ -36,9 +36,9 @@ import com.markoala.tomoandroid.data.repository.friends.FriendsRepository
 import com.markoala.tomoandroid.ui.components.CustomText
 import com.markoala.tomoandroid.ui.components.CustomTextType
 import com.markoala.tomoandroid.ui.components.DashedBorderBox
-import com.markoala.tomoandroid.ui.components.ErrorHandler
 import com.markoala.tomoandroid.ui.components.LocalToastManager
 import com.markoala.tomoandroid.ui.theme.CustomColor
+import com.markoala.tomoandroid.utils.ErrorHandler
 
 @Composable
 fun AddFriendsScreen(
