--- frontend/app/src/main/java/com/markoala/tomoandroid/utils/.gitkeep
@@ -1,2 +1,0 @@
-// 공통 Util을 위한 패키지입니다.
-
